Duplicates Refactor class added
