+++ src/main/java/employee_csv_project/model/EmployeesDAO.java
@@ -1,6 +1,6 @@
 package employee_csv_project.model;
 
-import employee_csv_project.controller.csv_controller.duplication_check.CheckForDuplicates;
+import employee_csv_project.controller.csv_controller.duplication_handler.CheckForDuplicates;
 
 import java.util.ArrayList;
 
@@ -10,10 +10,10 @@ public class EmployeesDAO {
     ArrayList<EmployeeDTO> allEmployees = new ArrayList<>();
 
     //places all employees into and ArrayList
-    public void addEmployeeToList(String[] employeeData){
-        EmployeeDTO employee = new EmployeeDTO(employeeData);
-        if(CheckForDuplicates.checkDuplicates(employee)) {
-            allEmployees.add(employee);
+    public void addEmployeeToList(EmployeeDTO employeeData){
+
+        if(CheckForDuplicates.checkDuplicates(employeeData)) {
+            allEmployees.add(employeeData);
         }
     }
